Rename IFAddress to InterfaceAddress
