+++ src/apple/network.rs
@@ -6,7 +6,8 @@ use std::collections::{hash_map, HashMap};
 use std::net::Ipv4Addr;
 use std::ptr::null_mut;
 
-use crate::socket::{MacAddr, get_interface_address, IFAddress};
+use crate::common::{InterfaceAddress, MacAddr};
+use crate::network::get_interface_address;
 use crate::{NetworkExt, NetworksExt, NetworksIter};
 
 macro_rules! old_and_new {
@@ -160,10 +161,10 @@ impl Networks {
             for (name, ifa) in iter {
                 if let Some(interface) = self.interfaces.get_mut(&name) {
                     match ifa {
-                        IFAddress::MAC(mac_addr) => {
+                        InterfaceAddress::MAC(mac_addr) => {
                             interface.mac_addr = mac_addr;
                         },
-                        IFAddress::IPv4(addr, mask) => {
+                        InterfaceAddress::IPv4(addr, mask) => {
                             interface.ipv4_addr = addr;
                             interface.ipv4_mask = mask;
                         }
